+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ in development
   [James Sigurðarson]
 * Allow user to filter executions on trigger instance id.
   [Sayli Karmarkar]
+* By default the following environment variables are now available to the actions executed by
+  local, remote and python runner: ``ST2_ACTION_API_URL``, ``ST2_ACTION_AUTH_TOKEN``. (new-feature)
 * Jinja filter to make working with regex and semver possible in any place that
   support jinja (improvement)